+++ inferia/__init__.py
@@ -1,9 +1,0 @@
-from inferia.core.app import Application
-from inferia.core.models import BasePredictor
-from inferia.core.utils import model_download
-
-__all__ = [
-    "Application",
-    "BasePredictor",
-    "model_download",
-]